backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
crewai==0.86.0
crewai-tools==0.17.0
langchain-google-genai==2.0.8
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
supabase==2.0.0
Faker==20.1.0
redis==5.0.0
openai>=1.13.3
mcp==1.1.0
neo4j==5.14.0
PyPDF2==3.0.1
google-generativeai==0.8.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
pandas==2.1.4
tabulate==0.9.0

